requirements.txt
Django
psycopg2-binary
django-webpack-loader
djangorestframework
scrapy
requests
google-cloud-bigquery
django-extensions
Pillow
pycocotools
django-background-tasks